NET-60: rename nodekeeper package to nodenetwork
