+++ network/dhtnetwork/dht.go
@@ -509,7 +509,7 @@ LOOP:
 
 func (dht *DHT) AddUnsync(nodeID core.RecordRef, roles []core.NodeRole, address string,
 	version string /*, publicKey *ecdsa.PublicKey*/) (chan core.Node, error) {
-	// TODO: return nodekeeper from helper method in HostHandler and remove this func and GetActiveNodes
+	// TODO: return nodenetwork from helper method in HostHandler and remove this func and GetActiveNodes
 	return dht.activeNodeKeeper.AddUnsync(nodeID, roles, address, version /*, publicKey*/)
 }
