+++ network/nodenetwork/nodekeeper_test.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package nodekeeper
+package nodenetwork
 
 import (
 	"sync"
