+++ network/nodenetwork/unsync_list.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package nodekeeper
+package nodenetwork
 
 import (
 	"errors"
