+++ testutils/nodekeeper/nodekeeper.go
@@ -3,14 +3,14 @@ package nodekeeper
 import (
 	"github.com/insolar/insolar/core"
 	"github.com/insolar/insolar/network"
-	"github.com/insolar/insolar/network/nodekeeper"
+	"github.com/insolar/insolar/network/nodenetwork"
 	testNetwork "github.com/insolar/insolar/testutils/network"
 )
 
 func GetTestNodekeeper() network.NodeKeeper {
 	nw := testNetwork.GetTestNetwork()
-	keeper := nodekeeper.NewNodeKeeper(
-		nodekeeper.NewNode(
+	keeper := nodenetwork.NewNodeKeeper(
+		nodenetwork.NewNode(
 			nw.GetNodeID(),
 			[]core.NodeRole{core.RoleVirtual, core.RoleHeavyMaterial, core.RoleLightMaterial},
 			nil,
@@ -23,7 +23,7 @@ func GetTestNodekeeper() network.NodeKeeper {
 
 	// dirty hack - we need 3 nodes as validators, pass one node 3 times
 	getValidator := func() core.Node {
-		return nodekeeper.NewNode(
+		return nodenetwork.NewNode(
 			nw.GetNodeID(),
 			[]core.NodeRole{core.RoleVirtual, core.RoleLightMaterial},
 			nil,
